renamed business_logic file
