+++ webviz_subsurface/plugins/_group_tree/_utils/__init__.py
@@ -1,1 +1,1 @@
-from ._business_logic import EnsembleGroupTreeData
+from ._ensemble_group_tree_data import EnsembleGroupTreeData
