+++ webviz_subsurface/plugins/_group_tree/_utils/_ensemble_group_tree_data.py
